requirements.txt
google-cloud-vision
google-cloud-language
pyrebase